fixing link error related to openwire classes
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@503343 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/connector/openwire/OpenwireCommandReader.h
@@ -82,7 +82,7 @@ namespace openwire{
          * @return The number of bytes read.
          * @throws IOException thrown if an error occurs.
          */
-        virtual int read( unsigned char* buffer, int count ) 
+        virtual size_t read( unsigned char* buffer, size_t count ) 
             throw( io::IOException );
        
         /**
